Muda tipagem e forma de lançar exceção para dominio de cartâo

+++ src/main/java/com/rafa/autorizador/cartao/saldo/BuscadorSaldo.java
@@ -1,7 +1,9 @@
 package com.rafa.autorizador.cartao.saldo;
 
+import java.math.BigDecimal;
+
 import com.rafa.autorizador.config.exception.CartaoInexistenteException;
 
 public interface BuscadorSaldo {
-    Double busca(String numeroCartao) throws CartaoInexistenteException;
+    BigDecimal busca(String numeroCartao) throws CartaoInexistenteException;
 }
